backend/cloud_functions/rag_preprocess/requirements.txt
functions-framework==3.*
google-cloud-storage>=2.0.0
google-cloud-tasks>=2.0.0
supabase>=1.0.0
python-docx>=1.0.0
# Add any other specific dependencies needed by the function
